Google form responses
This dataset contains the google form responses of the subjects that were provided to them as the final stage of data collection experiment. 
The columns C1 through C18 are questions corresponding to questions.png file and the score was calculated based on the perceived stress scale.pdf file.
The label then corresponds to the score obtained.